Create index.php
Wordpress plugin to allow setting website alternate link for your website. 
If you are targeting your website for a particular region then you can mention the language code and country code for the website using this plugin. 

After installing this plugin and activating it you will need to set the language code - country code on the Settings page

<link rel="alternate" href="https://www.adeelsarfaraz.com" hreflang="en-pk" />

You will need to enter en-pk through settings page of the plugin to accomplish this.